requirements.txt
openai
chromadb
python-dotenv
pandas
numpy
fastapi[all]
uvicorn